Structural Issue: The console consumes * characters, preventing simple math functions
The * character is used for diagnostic queries between the machine and the controller. It is not used by smoothieware and is a makera addition.

By consuming * characters, the firmware prevents us from implementing multiplication as 1*1 returns 2 in gcode, but 11 in the console. (in our dev branch),

If the character were changed to '&' instead, there would be no conflicts.

This would also require a change to controller.py line 718  self.stream.send(b"*") -> self.stream.send(b"&")
and 397