Front Button Manual Tool Change
front button manual tool changes test firmware

how it works: when the machine is idle or paused, press the front button. If the machine is unhomed or clamped, it will loosen the collet
if the tool is unclamped, it will tighen the collet

if the machine is running, it halt. That way the bit isn't dropped by a moving spindle

while it is moving, there is a popup in the controller telling you it is changing tools. Don't hit confirm, it will clear itself
to get the machine into toolchange front button mode you need to run
config-set sd main_button_long_press_enable ToolChange
in the MDI