Add zprobe.calibration_safety_margin
Since read_calibrate now detects whether you're using a probe, and
requires both the TLO sensor and probe sensor to activate before
halting, we need some protection against crashes.

This could most obviously occur if the probe is disconnected, but could
also occur if a NC probe is used since that case is not yet handled for
G38.6.

Since the read_calibrate changes are intended to account for minor
offsets between activation of the TLO sensor and the tool probe, in the
order of 10s of microns, this safety margin value is set to a
conservative value of 0.1mm.

This change can be tested by initiating a G38.6 Z-10 F50 probe, then
pushing the TLO sensor. The console should report an error that no probe
signal was deteced within 0.1mm of Z axis travel.