S2 only set's the Z offset of the current wcs if the last probing was ok, meaning the probe triggered.
User get's an error and machine halt when he tries to do an auto z probe starting a file, setting Z offset with G10 L20 or running a M46x macro with S2 parameter after he set the tool to a probing tool (0 or >= 999990) without calibrating the tlo. This "probe not calibrated state" is also written to the eeprom.

The TLO is cleared now and the reference tool mz value is set to the current tool mz when the Z wcs offset is set with a M46x macro.