Make check-version-txt more robust
check-version-txt.sh was falsely reporting success. Use a simpler method to find the changed base information, and log what is going on in case this fails again in the future.

Also fix up issues with the way checks are reported, they require a
different SHA to the one we were using.

Finally, add something for the summary page.